requirements.txt
black==22.12.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
idna==3.4
instagrapi==1.16.31
mypy-extensions==0.4.3
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
pycryptodomex==3.15.0
pydantic==1.10.2
PySocks==1.7.1
python-dotenv==0.21.1
requests==2.28.2
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.14
numpy
pandas
openai